[RHOL-NOTICKET]: Move ErrorLog class to separate file
