+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/Runtime.scala
@@ -32,7 +32,7 @@ class Runtime private (val reducer: Reduce[Task],
                                                      ListChannelWithRandom,
                                                      ListChannelWithRandom,
                                                      TaggedContinuation],
-                       var errorLog: Runtime.ErrorLog) {
+                       var errorLog: ErrorLog) {
   def readAndClearErrorVector(): Vector[InterpreterError] = errorLog.readAndClearErrorVector()
   def close(): Unit                                       = space.close()
 }
@@ -62,40 +62,6 @@ object Runtime {
           TaggedContinuation(ScalaBodyRef(ref))
         )
     }
-
-  class ErrorLog extends FunctorTell[Task, InterpreterError] {
-    private var errorVector: Vector[InterpreterError] = Vector.empty
-    val functor                                       = implicitly[Functor[Task]]
-    override def tell(e: InterpreterError): Task[Unit] =
-      Task.now {
-        this.synchronized {
-          errorVector = errorVector :+ e
-        }
-      }
-
-    override def writer[A](a: A, e: InterpreterError): Task[A] =
-      Task.now {
-        this.synchronized {
-          errorVector = errorVector :+ e
-        }
-        a
-      }
-
-    override def tuple[A](ta: (InterpreterError, A)): Task[A] =
-      Task.now {
-        this.synchronized {
-          errorVector = errorVector :+ ta._1
-        }
-        ta._2
-      }
-
-    def readAndClearErrorVector(): Vector[InterpreterError] =
-      this.synchronized {
-        val ret = errorVector
-        errorVector = Vector.empty
-        ret
-      }
-  }
 
   def create(dataDir: Path, mapSize: Long)(implicit captureTask: Capture[Task]): Runtime = {
