+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/ReduceSpec.scala
@@ -49,7 +49,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   implicit val rand: Blake2b512Random = Blake2b512Random(Array.empty[Byte])
 
   "evalExpr" should "handle simple addition" in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val result = withTestSpace { space =>
       val reducer      = RholangOnlyDispatcher.create[Task, Task.Par](space).reducer
       val addExpr      = EPlus(GInt(7), GInt(8))
@@ -64,7 +64,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "evalExpr" should "leave ground values alone" in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val result = withTestSpace { space =>
       val reducer      = RholangOnlyDispatcher.create[Task, Task.Par](space).reducer
       val groundExpr   = GInt(7)
@@ -79,7 +79,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "evalExpr" should "handle equality between arbitary processes" in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val result = withTestSpace { space =>
       val reducer      = RholangOnlyDispatcher.create[Task, Task.Par](space).reducer
       val eqExpr       = EEq(GPrivate("private_name"), GPrivate("private_name"))
@@ -93,7 +93,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "evalExpr" should "substitute before comparison." in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val result = withTestSpace { space =>
       val reducer           = RholangOnlyDispatcher.create[Task, Task.Par](space).reducer
       implicit val emptyEnv = Env.makeEnv(Par(), Par())
@@ -107,7 +107,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of Bundle" should "evaluate contents of bundle" in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand = rand.splitByte(0)
     val result = withTestSpace { space =>
       val reducer = RholangOnlyDispatcher.create[Task, Task.Par](space).reducer
@@ -139,7 +139,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   it should "throw an error if names are used against their polarity" in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     /* for (n <- @bundle+ { y } ) { }  -> for (n <- y) { }
      */
     val y = GString("y")
@@ -175,7 +175,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of Send" should "place something in the tuplespace." in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand = rand.splitByte(0)
     val result = withTestSpace { space =>
       val reducer = RholangOnlyDispatcher.create[Task, Task.Par](space).reducer
@@ -207,7 +207,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   it should "verify that Bundle is writeable before sending on Bundle " in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand = rand.splitByte(0)
     /* @bundle+ { x } !(7) -> x!(7)
      */
@@ -231,7 +231,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of single channel Receive" should "place something in the tuplespace." in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand = rand.splitByte(0)
     val result = withTestSpace { space =>
       val receive =
@@ -278,7 +278,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   it should "verify that bundle is readable if receiving on Bundle" in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand = rand.splitByte(1)
     /* for (@Nil <- @bundle- { y } ) { }  -> for (n <- y) { }
      */
@@ -317,7 +317,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of Send | Receive" should "meet in the tuplespace and proceed." in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand0 = rand.splitByte(0)
     val splitRand1 = rand.splitByte(1)
     val mergeRand = Blake2b512Random.merge(Seq(splitRand1, splitRand0))
@@ -374,7 +374,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
 
   "eval of Send | Receive" should "when whole list is bound to list remainder, meet in the tuplespace and proceed. (RHOL-422)" in {
     // for(@[...a] <- @"channel") { … } | @"channel"!([7,8,9])
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand0 = rand.splitByte(0)
     val splitRand1 = rand.splitByte(1)
     val mergeRand = Blake2b512Random.merge(Seq(splitRand1, splitRand0))
@@ -432,7 +432,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of Send on (7 + 8) | Receive on 15" should "meet in the tuplespace and proceed." in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand0 = rand.splitByte(0)
     val splitRand1 = rand.splitByte(1)
     val mergeRand = Blake2b512Random.merge(Seq(splitRand1, splitRand0))
@@ -488,7 +488,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of Send of Receive | Receive" should "meet in the tuplespace and proceed." in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val baseRand = rand.splitByte(2)
     val splitRand0 = baseRand.splitByte(0)
     val splitRand1 = baseRand.splitByte(1)
@@ -592,7 +592,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "Simple match" should "capture and add to the environment." in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand = rand.splitByte(0)
     val result = withTestSpace { space =>
       val pattern = Send(ChanVar(FreeVar(0)), List(GInt(7), EVar(FreeVar(1))), false, BitSet())
@@ -639,7 +639,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of Send | Send | Receive join" should "meet in the tuplespace and proceed." in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand0 = rand.splitByte(0)
     val splitRand1 = rand.splitByte(1)
     val splitRand2 = rand.splitByte(2)
@@ -721,7 +721,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of Send with remainder receive" should "capture the remainder." in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand0 = rand.splitByte(0)
     val splitRand1 = rand.splitByte(1)
     val mergeRand = Blake2b512Random.merge(Seq(splitRand1, splitRand0))
@@ -757,7 +757,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of nth method" should "pick out the nth item from a list" in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand = rand.splitByte(0)
     val nthCall: Expr =
       EMethod("nth", EList(List(GInt(7), GInt(8), GInt(9), GInt(10))), List[Par](GInt(2)))
@@ -800,7 +800,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of nth method in send position" should "change what is sent" in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand = rand.splitByte(0)
     val nthCallEvalToSend: Expr =
       EMethod("nth",
@@ -841,7 +841,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of `toByteArray` method on any process" should "return that process serialized" in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRand = rand.splitByte(0)
     import coop.rchain.models.serialization.implicits._
     val proc = Receive(Seq(ReceiveBind(Seq(ChanVar(FreeVar(0))), Quote(GString("channel")))),
@@ -877,7 +877,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   it should "return an error when `toByteArray` is called with arguments" in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val toByteArrayWithArgumentsCall: EMethod =
       EMethod(
         "toByteArray",
@@ -900,7 +900,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
 
   "eval of hexToBytes" should "transform encoded string to byte array (not the rholang term)" in {
     import coop.rchain.models.serialization.implicits._
-    implicit val errorLog         = new Runtime.ErrorLog()
+    implicit val errorLog         = new ErrorLog()
     val splitRand = rand.splitByte(0)
     val testString                = "testing testing"
     val base16Repr                = Base16.encode(testString.getBytes)
@@ -933,7 +933,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "variable references" should "be substituted before being used." in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRandResult = rand.splitByte(3)
     val splitRandSrc = rand.splitByte(3)
     splitRandResult.next()
@@ -958,7 +958,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
     )
 
     val result = withTestSpace { space =>
-      implicit val errorLog = new Runtime.ErrorLog()
+      implicit val errorLog = new ErrorLog()
       val reducer           = RholangOnlyDispatcher.create[Task, Task.Par](space).reducer
       val env               = Env[Par]()
       val task              = reducer.eval(proc)(env, splitRandSrc)
@@ -980,7 +980,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   it should "be substituted before being used in a match." in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val splitRandResult = rand.splitByte(4)
     val splitRandSrc = rand.splitByte(4)
     splitRandResult.next()
@@ -995,7 +995,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
     )
 
     val result = withTestSpace { space =>
-      implicit val errorLog = new Runtime.ErrorLog()
+      implicit val errorLog = new ErrorLog()
       val reducer           = RholangOnlyDispatcher.create[Task, Task.Par](space).reducer
       val env               = Env[Par]()
       val task              = reducer.eval(proc)(env, splitRandSrc)
@@ -1017,7 +1017,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   it should "reference a variable that comes from a match in tuplespace" in {
-    implicit val errorLog = new Runtime.ErrorLog()
+    implicit val errorLog = new ErrorLog()
     val baseRand = rand.splitByte(7)
     val splitRand0 = baseRand.splitByte(0)
     val splitRand1 = baseRand.splitByte(1)
@@ -1045,7 +1045,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
     )
 
     val result = withTestSpace { space =>
-      implicit val errorLog = new Runtime.ErrorLog()
+      implicit val errorLog = new ErrorLog()
       val reducer           = RholangOnlyDispatcher.create[Task, Task.Par](space).reducer
       val env               = Env[Par]()
       val task              = reducer.eval(proc)(env, baseRand)
